- fix negativ cpu time reported [windows]  - fix no fraction done reported when using    "boinc fraction_done"
 - small cleanup



git-svn-id: svn+ssh://cvs.lpds.sztaki.hu/var/lib/svn/szdg/apps/genwrapper/trunk@2093 a7169a2c-3604-0410-bc95-c702d8d87f7a

+++ launcher/task.C
@@ -347,7 +347,7 @@ bool TASK::poll(int& status) {
       gw_do_log(LOG_ERR, "failed to query information on JobObject (%ld)", 
 		(long)GetLastError());
     } else {
-      final_cpu_time = ((long)Rusage.TotalUserTime.QuadPart) / 10000000;
+      final_cpu_time = (double)(((long long)Rusage.TotalUserTime.QuadPart) / 10000000);
     }
     gw_report_status(final_cpu_time, frac_done, false);
     if (exit_code != STILL_ACTIVE) {
